refactor(config): ♻️ merge options flow into config_flow.py
- move NeoPoolOptionsFlowHandler from options_flow.py into config_flow.py
- delete the now-empty options_flow.py module
- merge test_options_flow.py into test_config_flow.py
- mirror the source-file layout so the options flow lives beside the config flow
